fix: resolve all compiler warnings and linker error across codebase
- Add -lm to off_server link libraries (undefined reference to round)
- Fix -Wunused-parameter with (void) casts in block, http, network,
  streams, refcounter source files
- Fix -Wunused-variable by removing or casting unused locals
- Fix -Wsign-compare with proper type casts (size_t vs int, uint64_t vs
  int32_t) across index, fibonacci, block_cache, wal, get_dir
- Fix -Wmissing-field-initializers with #pragma GCC diagnostic
  push/pop around hashmap_foreach_data macro expansions
- Fix -Wincompatible-pointer-types in ofd_cache hashmap typedef
- Fix -Wunused-result with (void) casts on cbor_array_push/cbor_map_add
- Fix -Wmacro-redefined with #ifndef guards on _GNU_SOURCE defines
- Fix -Wwritable-strings with (char*) casts on string literal paths
- Fix -Wmissing-braces with memset pattern in test files
- Add __attribute__((unused)) to static functions not yet called

+++ src/BlockCache/wal.c
@@ -94,7 +94,7 @@ int wal_read(wal_t* wal, wal_type_e* type, buffer_t** data, uint64_t* cursor, in
     if (lseek(wal->log, 0, SEEK_SET) < 0) {
       return -2;
     }
-  } else if (*cursor >= *wal_size) {
+  } else if (*cursor >= (uint64_t)*wal_size) {
       return -3;
   }
   lseek(wal->log, *cursor, SEEK_SET);
