+++ src/Util/get_dir.c
@@ -42,7 +42,7 @@ void destroy_files(vec_str_t* files) {
   if (files == NULL) {
     return;
   }
-  size_t i; char* file;
+  int i; char* file;
   vec_foreach(files, file, i) {
       free(file);
   }
